requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
amqp==5.1.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
billiard==4.1.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
exceptiongroup==1.1.1
filelock==3.12.0
Flask==2.2.2
frozenlist==1.3.3
fsspec==2023.5.0
gunicorn==20.1.0
h11==0.14.0
huggingface-hub==0.15.1
idna==3.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kombu==5.3.1
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
openai==0.26.4
outcome==1.2.0
packaging==23.0
pipeline==0.1.0
prompt-toolkit==3.0.38
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
redis==4.5.4
regex==2023.6.3
requests==2.28.2
safetensors==0.3.1
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tokenizers==0.13.3
tqdm==4.64.1
transformers==4.30.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.14
vine==5.0.0
waitress==2.1.2
wcwidth==0.2.6
webdriver-manager==3.8.5
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
